requirements.txt
annotated-types==0.7.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
cryptography==42.0.8
diskcache==5.6.3
future==1.0.0
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
junos-eznc==2.7.1
lxml==5.2.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.3.0
netutils==1.8.1
nornir==3.4.1
nornir-netmiko==1.0.1
nornir-salt==0.21.1
nornir-table-inventory==0.4.5
nornir-utils==0.2.0
nornir_napalm==0.5.0
ntc_templates==5.1.0
numpy==1.26.4
pandas==1.5.3
paramiko==3.4.0
picle==0.5.3
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pyeapi==1.0.2
PyNaCl==1.5.0
pyparsing==3.1.2
pyreadline3==3.4.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.15.0
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.30
textfsm==1.1.3
transitions==0.9.1
ttp==0.9.5
ttp-templates==0.3.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
yamlordereddictloader==0.4.2
